# core/api/docs/v2/openapi.yaml
components:
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Invalid path, query or request input
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
      description: Missing, unknown, revoked or expired key; shared authentication envelope
    Forbidden:
      content:
        application/json:
          schema:
            anyOf:
            - $ref: '#/components/schemas/ForbiddenError'
            - $ref: '#/components/schemas/Error'
      description: The shared key-scope gate uses the legacy envelope; an operation-specific refusal may use the v2 error envelope
    Conflict:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Idempotency-Key conflict, or an operation-specific concurrency conflict
...
    RateLimited:
      content:
        application/json:
          schema:
            properties:
              code:
                example: rate_limit_exceeded
                type: string
              message:
                example: Rate limit exceeded
                type: string
              object:
                example: error
                type: string
              status:
                example: 429
                type: integer
            type: object
      description: Shared write rate limit; legacy error envelope
  schemas:
    ForbiddenError:
      properties:
        code:
          example: forbidden
          type: string
        message:
          example: Forbidden
          type: string
        object:
          example: error
          type: string
        status:
          example: 403
          type: integer
      type: object
    UnauthorizedError:
      properties:
        code:
          example: unauthorized
          type: string
        message:
          example: Unauthorized
          type: string
        object:
          example: error
          type: string
        status:
          example: 401
          type: integer
      type: object
...
    CreateResult:
      properties:
        created:
          $ref: '#/components/schemas/SideEffects'
        dry_run:
          type: boolean
        etag:
          description: etag of the created object
          type: string
        id:
          type: string
        issues:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
        key:
          description: identity key (types, properties)
          type: string
        type:
          description: type key of the created object
          type: string
        warnings:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
      type: object
...
    CreatedOption:
      properties:
        name:
          type: string
        property:
          description: property key
          type: string
      type: object
    DiffStats:
      properties:
        blocks_added:
          type: integer
        blocks_changed:
          type: integer
        blocks_moved:
          type: integer
        blocks_removed:
          type: integer
        properties_changed:
          type: integer
      type: object
...
    EditResult:
      properties:
        created:
          $ref: '#/components/schemas/SideEffects'
        created_blocks:
          additionalProperties:
            type: string
          description: |-
            CreatedBlocks maps each payload position that created a block to the
            id the server minted for it: the top-level run positions
            ("ops[3].blocks[0]") and the nested slots alike, such as a table's
            rows and columns ("ops[3].blocks[0].rows[1]") and the blocks inside a
            cell run ("ops[3].value[1]"). A position that carried an id is
            absent, because the block it names already existed.
          type: object
        created_views:
          additionalProperties:
            type: string
          description: |-
            CreatedViews maps each payload position that created a dataview view
            to the view id the server minted: an insert_view op ("ops[i]"), or a
            view slot of an update_block set channel ("ops[i].set.views[2]").
            View ids are always server-minted, and a view is not a block, so they
            are reported here rather than in CreatedBlocks.
          type: object
        diff_stats:
          $ref: '#/components/schemas/DiffStats'
        dry_run:
          type: boolean
        etag:
          type: string
        warnings:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
      type: object
    Error:
      properties:
        code:
          type: string
        issues:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
        message:
          type: string
        status:
          type: integer
      type: object
...
    Issue:
      properties:
        hint:
          type: string
        message:
          type: string
        path:
          type: string
      type: object
...
    ListResponse-ObjectRow:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ObjectRow'
          type: array
          uniqueItems: false
        has_more:
          type: boolean
        limit:
          type: integer
        message:
          type: string
        offset:
          type: integer
        total:
          type: integer
        warnings:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
      type: object
...
    ObjectRow:
      properties:
        id:
          type: string
        name:
          type: string
        properties:
          additionalProperties: {}
          type: object
        space_id:
          type: string
        type:
          type: string
      type: object
...
    PropertyRow:
      properties:
        format:
          type: string
        key:
          type: string
        name:
          type: string
      type: object
...
    SideEffects:
      properties:
        options:
          items:
            $ref: '#/components/schemas/CreatedOption'
          type: array
          uniqueItems: false
        properties:
          items:
            $ref: '#/components/schemas/PropertyRow'
          type: array
          uniqueItems: false
      type: object
...
    ValidateResponse:
      properties:
        issues:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
        warnings:
          items:
            $ref: '#/components/schemas/Issue'
          type: array
          uniqueItems: false
      type: object
...
  securitySchemes:
    bearerauth:
      scheme: bearer
      type: http
...
info:
  contact:
    email: support@anytype.io
    name: Anytype Support
    url: https://anytype.io/contact
  description: |-
    The agent-oriented Anytype local API. An object is one AnyBlock JSON document rather than a tree of blocks, so a single GET returns a whole editable document and a single PATCH edits it.
    Everything this API names for itself is snake_case: path and query parameters, request and response fields, and the names of the PATCH ops. Things are addressed by name rather than by id: property keys, option names, and `type` as a type key. Object ids remain ids.
    Property and type keys have two spellings, and a served body picks one. The default is the key: a snake_case slug minted when the property was created and frozen from then on, so it survives every rename. Store that spelling in a config, a script, an integration. `?keys=name` on a read serves the display-name vocabulary instead, "Due date" rather than due_date, which is the spelling to show a user their own words. Input needs no such choice: every channel accepts both, plus the forgiving fold between them, so dueDate, due-date, "Due date" and Дата выполнения all address the property they name. A spelling matching two properties is a 400 listing both, and an unknown one never creates anything.
...
    A read never fails on content it cannot represent. Whatever a representation cannot express is reported in `warnings` beside the result.
...
  license:
    name: Any Source Available License 1.0
...
openapi: 3.1.0
paths:
...
  /v2/spaces/{space_id}/objects:
    get:
      operationId: list_objects
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Comma-separated property keys to include per row
        in: query
        name: fields
        schema:
          type: string
      - description: Items to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Items to return
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-ObjectRow'
          description: Minimal object rows
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid fields query
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found or unavailable
      security:
      - bearerauth: []
      summary: List the objects in a space
      tags:
      - Objects
    post:
      description: A select value naming an option the property does not hold is refused
        unless `create_missing_options=true` is set. An unknown type or property key
        is rejected either way, with the closest matches named. The body is either
        a full AnyBlock document or the shortcut {type, name, properties, markdown};
        `formatVersion` or `blocks` picks the document form.
      operationId: create_object
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: 'Create select options for names the property does not hold yet
          (default false: an unmatched name is refused)'
        in: query
        name: create_missing_options
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Created object id + etag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation or reference failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create an object
      tags:
      - Objects
  /v2/spaces/{space_id}/objects/{object_id}:
    delete:
      description: Only objects this key created can be deleted. The creator is recorded
        at creation time and never added later, so objects made in the app, imported,
        made by another member, or made before this route shipped are refused for
        good. System objects are a 403 as well. A dry run reports the verdict without
        the checks that run at archive time, so a deletable verdict can still meet
        a 403.
      operationId: delete_object
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Object id
        in: path
        name: object_id
        required: true
        schema:
          type: string
      - description: Probe deletability without writing
        in: query
        name: dry_run
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Archived object, or the dry-run verdict. Deleting again is
            a 200 carrying a warning.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: 'A type or a property: use their own delete routes'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Object or space not found
        "409":
          $ref: '#/components/responses/Conflict'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Delete an object this key created
      tags:
      - Objects
    get:
      description: A `block` subtree comes back flagged as a subtree, and no write
        path accepts that partial body. `format=md` is read-only; markdown cannot
        be sent back.
      operationId: get_object
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Object id
        in: path
        name: object_id
        required: true
        schema:
          type: string
      - description: Subset of properties,blocks (default both)
        in: query
        name: include
        schema:
          type: string
      - description: 'Return the outline instead of full blocks: every block''s indent/id/type
          plus its text truncated to 80 runes'
        in: query
        name: outline
        schema:
          type: boolean
      - description: Return only this block's subtree
        in: query
        name: block
        schema:
          type: string
      - description: compact (default) is the edit shape, where minted block ids relabel
          to short suffixes; full is the export shape, with full ids everywhere, and
          the shape to send back. Object references are full and inline in both.
        in: query
        name: ids
        schema:
          type: string
      - description: anyblock (default) or md
        in: query
        name: format
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: The flat AnyBlock document + etag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Illegal parameter combination (ambiguous_input)
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Object or space not found
      security:
      - bearerauth: []
      summary: Read an object as an AnyBlock document
      tags:
      - Objects
    patch:
      description: Ops apply in order as one change set. If one fails, or the result
        breaks the format's rules, none of them land. `update_block`, `delete_block`
        and `replace_text` can address a block by its exact text instead of an id;
        text matching zero or several blocks is refused, not guessed at. A later op
        sees the earlier ones' edits. Ops that only create take no id; the new ids
        come back in `created_blocks`.
      operationId: patch_object
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Object id
        in: path
        name: object_id
        required: true
        schema:
          type: string
      - description: The etag the object must still carry
        in: header
        name: If-Match
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditResult'
          description: New etag + created block ids + diff_stats
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid ops or post-op document
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Object, space, or referenced block not found
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Existing content cannot be represented losslessly for this
            edit
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Edit an object with a batch of ops
      tags:
      - Objects
...
  /v2/validate:
    post:
      description: 'Structure and format rules only. Nothing is resolved against a
        space, so option names and a type''s property keys are not checked here. Findings
        come back as data: an invalid document is still a 200, carrying the issues,
        and a valid one carries empty lists.'
      operationId: validate
      parameters:
      - description: Retry key; identical requests replay, different requests conflict
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
          description: Issue and warning lists, empty when the document is valid
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
      security:
      - bearerauth: []
      summary: Validate an AnyBlock document
      tags:
      - Schemas
servers:
...